+++ misc/scripts/fetch-sources.sh
@@ -22,13 +22,16 @@
 # You should have received a copy of the GNU General Public License
 # along with Pacstall. If not, see <https://www.gnu.org/licenses/>.
 
+{ ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
+
 # shellcheck source=./misc/scripts/build.sh
 source "${SCRIPTDIR}/scripts/build.sh" || {
     fancy_message error "Could not find build.sh"
-    return 1
+    { ignore_stack=true; return 1; }
 }
 
 function parse_source_entry() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     unset source_url dest git_branch git_tag git_commit
     local entry="$1"
     source_url="${entry#*::}"
@@ -64,6 +67,7 @@ function parse_source_entry() {
 }
 
 function calc_git_pkgver() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     unset comp_git_pkgver
     local calc_commit
     if [[ $source_url == git+* ]]; then
@@ -82,6 +86,7 @@ function calc_git_pkgver() {
 }
 
 function genextr_declare() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     unset ext_method ext_deps
     # shellcheck disable=SC2031,SC2034
     case "${source_url,,}" in
@@ -149,12 +154,14 @@ function genextr_declare() {
 }
 
 function clean_fail_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     fancy_message info "Cleaning up"
     cleanup
     exit 1
 }
 
 function hashcheck() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local inputFile="${1}" inputHash="${2}" hashMethod="${3}sum" fileHash
     # Get hash of file
     fileHash="$(${hashMethod} "${inputFile}")"
@@ -172,12 +179,14 @@ function hashcheck() {
 }
 
 function fail_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     error_log 1 "download ${pacname}"
     fancy_message error "Failed to download package"
     clean_fail_down
 }
 
 function gather_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     export srcdir="${PACDIR}/${pacname}~${pkgver}"
     mkdir -p "${srcdir}"
     cd "${srcdir}" || {
@@ -188,6 +197,7 @@ function gather_down() {
 }
 
 function git_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local revision gitopts submodules=true no_submodule silence quiet
     ${PACSTALL_VERBOSE} || silence=("&>" "/dev/null") quiet="--quiet"
     dest="${dest%.git}"
@@ -252,19 +262,22 @@ function git_down() {
 }
 
 function net_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     fancy_message info "Downloading ${BPurple}${dest}${NC}"
     # shellcheck disable=SC2031
     download "$source_url" "$dest" || fail_down
 }
 
 function hashcheck_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     if [[ -n ${expectedHash} && ${expectedHash} != "SKIP" ]]; then
         fancy_message sub "Checking hash ${YELLOW}${expectedHash:0:8}${NC}[${YELLOW}...${NC}]"
-        hashcheck "${dest}" "${expectedHash}" "${hashsum_method}" || return 1
+        hashcheck "${dest}" "${expectedHash}" "${hashsum_method}" || { ignore_stack=true; return 1; }
     fi
 }
 
 function genextr_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     hashcheck_down
     local extract=true keep_archive
     for keep_archive in "${noextract[@]}"; do
@@ -277,7 +290,7 @@ function genextr_down() {
         fancy_message sub "Extracting ${CYAN}${dest}${NC}"
         ${ext_method} "${dest}" 1>&1 2> /dev/null
         if [[ -f ${dest} ]]; then
-            rm -f "${dest}"
+            rm -f "${dest:?}"
         fi
     fi
     # if first source and extract is true, enter it for archive check
@@ -297,6 +310,7 @@ function genextr_down() {
 }
 
 function deb_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     hashcheck_down
     local upgrade=false
     if is_package_installed "${pacname}" && type -t pre_upgrade &> /dev/null; then
@@ -361,6 +375,7 @@ function deb_down() {
 }
 
 function file_down() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     fancy_message info "Copying local archive ${BPurple}${dest}${NC}"
     # shellcheck disable=SC2031
     cp -r "${source_url}" "${dest}" || fail_down
@@ -392,6 +407,7 @@ function file_down() {
 
 # currently expecting: 1=hash 2=PACSTALL_KNOWN_SUMS 3=hashum_method 4=${CARCH}/${DISTRO} 5=${CARCH}
 function append_hash_entry() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local -n append="${1}" sums="${2}" exp_method="${3}"
     local hash_arch hash_arr extend="${4}${5:+_$5}"
     for type in "${sums[@]}"; do
@@ -416,6 +432,7 @@ function append_hash_entry() {
 }
 
 function append_var_arch() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local inp inputvar="${1}" inputvar_arch="${1}_${2}${3:+_$3}[*]"
     declare -n ref_inputvar="${inputvar}"
     if [[ -n ${!inputvar_arch} ]]; then
@@ -430,6 +447,7 @@ function append_var_arch() {
 }
 
 function append_modifier_entries() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     unset hashsum_method
     # shellcheck disable=SC2034
     local APPARCH="${1}" APPDISTRO="${2}"
@@ -461,6 +479,7 @@ function append_modifier_entries() {
 }
 
 function calc_distro() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local distro_pretty_name key value
     while IFS='=' read -r key value; do
         case "${key}" in
@@ -491,6 +510,7 @@ function calc_distro() {
 }
 
 function set_distro() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local distro_name distro_version_name distro_version_number distro_parent distro_parent_vname distro_parent_number
     calc_distro
     if [[ ${1} == "parent" ]]; then
@@ -501,6 +521,7 @@ function set_distro() {
 }
 
 function get_compatible_releases() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     # example for this function is "ubuntu:jammy"
     local distro_name distro_version_name distro_version_number distro_parent distro_parent_vname distro_parent_number is_compat=false comp_list=("${@,,}")
     calc_distro
@@ -533,12 +554,13 @@ function get_compatible_releases() {
     done
     if [[ ${is_compat} == "false" || ${is_compat} != "true" ]]; then
         fancy_message error "This Pacscript does not work on ${BBlue}${distro_name}:${distro_version_name}${NC}/${BBlue}${distro_name}:${distro_version_number}${NC}"
-        return 1
+        { ignore_stack=true; return 1; }
     fi
     return 0
 }
 
 function get_incompatible_releases() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     # example for this function is "ubuntu:jammy"
     local distro_name distro_version_name distro_version_number distro_parent distro_parent_vname distro_parent_number incomp_list=("${@,,}")
     calc_distro
@@ -566,26 +588,27 @@ function get_incompatible_releases() {
             # check for `22.04` or `jammy`
             if [[ ${key#*:} == "${distro_version_number}" || ${key#*:} == "${distro_version_name}" ]]; then
                 fancy_message error "This Pacscript does not work on ${BBlue}${distro_version_name}${NC}/${BBlue}${distro_version_number}${NC}"
-                return 1
+                { ignore_stack=true; return 1; }
             fi
         # check for `ubuntu:*`
         elif [[ $key == *":*" ]]; then
             # check for `ubuntu`
             if [[ ${key%%:*} == "${distro_name}" ]]; then
                 fancy_message error "This Pacscript does not work on ${BBlue}${distro_name}${NC}"
-                return 1
+                { ignore_stack=true; return 1; }
             fi
         else
             # check for `ubuntu:jammy` or `ubuntu:22.04`
             if [[ $key == "${distro_name}:${distro_version_name}" || $key == "${distro_name}:${distro_version_number}" ]]; then
                 fancy_message error "This Pacscript does not work on ${BBlue}${distro_name}:${distro_version_name}${NC}/${BBlue}${distro_name}:${distro_version_number}${NC}"
-                return 1
+                { ignore_stack=true; return 1; }
             fi
         fi
     done
 }
 
 function is_compatible_arch() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local inarch=("${@}") ret=1 pacarch farch
     # shellcheck disable=SC2076,SC2153
     if array.contains inarch "any" \
@@ -611,10 +634,11 @@ function is_compatible_arch() {
     if ((ret == 1)); then
         fancy_message error "This Pacscript does not work on ${BBlue}${CARCH}/${AARCH}${NC}"
     fi
-    return "${ret}"
+    { ignore_stack=true; return "${ret}"; }
 }
 
 function install_builddepends() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     # shellcheck disable=SC2034
     local build_dep not_installed_yet_builddepends bdeps_array bdeps_str check_dep not_installed_yet_checkdepends cdeps_array bcons_array bcons_str
     if [[ -n ${makedepends[*]} ]]; then
@@ -674,9 +698,11 @@ function install_builddepends() {
 }
 
 function compare_remote_version() {
+    { ignore_stack=false; set -o pipefail; trap stacktrace ERR RETURN; }
     local crv_input="${1}" remote_tmp remote_safe remotever localver crv_pkgver crv_pkgrel crv_epoch crv_source remv crv_fetch crv_base
     unset _pkgbase
-    source "$METADIR/$crv_input" || return 1
+    # shellcheck source=/dev/null
+    source "$METADIR/$crv_input" || { ignore_stack=true; return 1; }
     [[ ${_remoterepo} == "orphan" ]] && _remoterepo="${REPO}"
     if [[ -z ${_remoterepo} ]]; then
         return 0
@@ -696,7 +722,8 @@ function compare_remote_version() {
         unset pkgrel
         remote_tmp="$(sudo mktemp -p "${PACDIR}" -t "compare-repo-ver-$crv_input.XXXXXX")"
         remote_safe="${remote_tmp}"
-        curl -fsSL "$remoterepo/packages/$crv_fetch/.SRCINFO" | sudo tee "${remote_safe}" > /dev/null || return 1
+        # shellcheck disable=SC2034
+        curl -fsSL "$remoterepo/packages/$crv_fetch/.SRCINFO" | sudo tee "${remote_safe}" > /dev/null || { ignore_stack=true; return 1; }
         sudo chown "${PACSTALL_USER}" "${remote_safe}"
         crv_base="$(srcinfo.match_pkg "${remote_safe}" pkgbase)"
         for remv in "pkgver" "pkgrel" "epoch"; do
@@ -712,7 +739,7 @@ function compare_remote_version() {
         else
             echo "${crv_epoch:+$crv_epoch:}${crv_pkgver}-pacstall${crv_pkgrel:-1}"
         fi
-        sudo rm -rf "${remote_safe}"
+        sudo rm -rf "${remote_safe:?}"
     )" > /dev/null
     localver=$(source "${METADIR}/${crv_input}" && echo "${_version}")
     if [[ $crv_input == *"-git" ]]; then
